Delete images from storage

+++ controllers/admin.js
@@ -1,5 +1,6 @@
 const Product = require('../models/product');
 // const mongodb = require('mongodb');
+const fileHelper = require('../util/file');
 
 const { validationResult } = require('express-validator');
 
@@ -167,6 +168,7 @@ exports.postEditProduct = (req, res, next) => {
             product.description = updatedDescription;
             //update if we give new image file
             if (image) {
+                fileHelper.deleteFile(product.imageUrl);
                 product.imageUrl = image.path;
             }
 
@@ -186,7 +188,15 @@ exports.postEditProduct = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({ _id: prodId, userId: req.user._id })
+
+    Product.findById(prodId)
+        .then(product => {
+            if (!product) {
+                return next(new Error('Product not found!'));
+            }
+            fileHelper.deleteFile(product.imageUrl);
+            Product.deleteOne({ _id: prodId, userId: req.user._id });
+        })
         .then(result => {
             console.log('DESTROYED PRODUCT');
             res.redirect('/admin/products');
